Fix uncontrolled input warning
`state.answers` must always be initialized to an array of empty strings
with length === `props.questions.length`. If not, the inputs will have
an undefined value, causing React to treat them as uncontrolled
inputs. Then later, when we try to set a value, React will warn that
we are trying to change an uncontrolled input into a controlled input.